test that is_guard changes when armedmen does.
