+++ src/guard.c
@@ -45,6 +45,20 @@ attrib_type at_guard = {
     ATF_UNIQUE
 };
 
+guard_t can_start_guarding(const unit * u)
+{
+    if (u->status >= ST_FLEE || fval(u, UFL_FLEEING))
+        return E_GUARD_FLEEING;
+    /* Monster der Monsterpartei dürfen immer bewachen */
+    if (is_monsters(u->faction) || fval(u_race(u), RCF_UNARMEDGUARD))
+        return E_GUARD_OK;
+    if (!armedmen(u, true))
+        return E_GUARD_UNARMED;
+    if (IsImmune(u->faction))
+        return E_GUARD_NEWBIE;
+    return E_GUARD_OK;
+}
+
 void update_guards(void)
 {
     const region *r;
