+++ src/laws.c
@@ -2671,20 +2671,6 @@ int combatspell_cmd(unit * u, struct order *ord)
     return 0;
 }
 
-guard_t can_start_guarding(const unit * u)
-{
-    if (u->status >= ST_FLEE || fval(u, UFL_FLEEING))
-        return E_GUARD_FLEEING;
-    /* Monster der Monsterpartei dürfen immer bewachen */
-    if (is_monsters(u->faction) || fval(u_race(u), RCF_UNARMEDGUARD))
-        return E_GUARD_OK;
-    if (!armedmen(u, true))
-        return E_GUARD_UNARMED;
-    if (IsImmune(u->faction))
-        return E_GUARD_NEWBIE;
-    return E_GUARD_OK;
-}
-
 int guard_on_cmd(unit * u, struct order *ord)
 {
     assert(getkeyword(ord) == K_GUARD);
